Initialise pd_this on audio thread
Since pd_this is thread local, it is generally only initialised on the
main thread. This causes an issue where the audio thread expects pd_this
to be non-null in order for sys_lock to succeed. This commit rectifies
the issue by setting pd_this in the audio callback.

Note that there may be other thread local state that needs to be
initialised. This has not been considered in this commit.