Avoid arithmetic error when buffer->size == 0
It is possible that a buffer is created with a zero size. In this case,
there is an arithmetic error trying to perform

`(buffer->size + write_idx - read_idx) % 0`

Really, the buffer should not be created with a zero size in the first
place, but this is a suitable fix for now.